requirements.txt
crewai
streamlit
python-dotenv